# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RoomInputMode = typeof RoomInputMode[keyof typeof RoomInputMode];


export const RoomInputMode = {
  NUMBER_7: 7,
  NUMBER_14: 14,
} as const;

export interface RoomPlayerInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  playerId: string;
  /**
     * @minLength 2
     * @maxLength 24
     */
  nickname: string;
  /** @maxLength 120 */
  avatar?: string;
}

export interface RoomInput {
  mode: RoomInputMode;
  host: RoomPlayerInput;
}

export interface RoomPlayer {
  playerId: string;
  nickname: string;
  /** @nullable */
  avatar: string | null;
  isHost: boolean;
  isReady: boolean;
}

export type RoomMode = typeof RoomMode[keyof typeof RoomMode];


export const RoomMode = {
  NUMBER_7: 7,
  NUMBER_14: 14,
} as const;

export type RoomStatus = typeof RoomStatus[keyof typeof RoomStatus];


export const RoomStatus = {
  LOBBY: 'LOBBY',
  STARTED: 'STARTED',
} as const;

export interface Room {
  code: string;
  mode: RoomMode;
  capacity: number;
  status: RoomStatus;
  players: RoomPlayer[];
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}