# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomMode } from './roomMode';
import type { RoomPlayer } from './roomPlayer';
import type { RoomStatus } from './roomStatus';

export interface Room {
  code: string;
  mode: RoomMode;
  capacity: number;
  status: RoomStatus;
  players: RoomPlayer[];
  createdAt: Date;
}