# lib/api-zod/src/generated/types/roomInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoomInputMode } from './roomInputMode';
import type { RoomPlayerInput } from './roomPlayerInput';

export interface RoomInput {
  mode: RoomInputMode;
  host: RoomPlayerInput;
}