# lib/api-zod/src/generated/types/roomPlayer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomPlayer {
  playerId: string;
  nickname: string;
  /** @nullable */
  avatar: string | null;
  isHost: boolean;
  isReady: boolean;
}