# lib/api-zod/src/generated/types/roomPlayerInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoomPlayerInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  playerId: string;
  /**
     * @minLength 2
     * @maxLength 24
     */
  nickname: string;
  /** @maxLength 120 */
  avatar?: string;
}